Add unregister notification command
This patch adds unreg-notif command. It is used to unregister the
previous callback that was set by reg-notif command. This way, the stack
will ignore any notification/indication received.

Note: This command does not disable notification/indication in remote
device. For that, you must write 0x00 byte in the respective client
characteristic configuration.